+++ bolt/utils/__init__.py
@@ -1,6 +1,7 @@
 from . import types
 from . import version
 from ._config import Config, CONFIG_PATH
+from ._download import download
 from ._yearmonth import YearMonth
 
 
@@ -11,6 +12,7 @@ __all__ = [
     CRS,
     config,
     CONFIG_PATH,
+    download,
     types,
     # ...
     version,